Collapse EnumStoreBase into EnumStoreT in preparation for replacing underlying data store with unique store.
