+++ searchlib/src/vespa/searchlib/attribute/loadedenumvalue.h
@@ -3,7 +3,8 @@
 #pragma once
 
 #include <vespa/vespalib/util/array.h>
-#include <vespa/searchlib/attribute/enumstorebase.h>
+#include <vespa/vespalib/util/arrayref.h>
+#include <vespa/searchlib/attribute/i_enum_store.h>
 
 namespace search
 {
