+++ searchlib/src/vespa/searchlib/attribute/loadedvalue.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include <vespa/searchcommon/common/undefinedvalues.h>
-#include <vespa/searchlib/attribute/enumstorebase.h>
+#include <vespa/searchlib/attribute/i_enum_store.h>
 #include <vespa/vespalib/datastore/entryref.h>
 
 namespace search
