+++ searchlib/src/vespa/searchlib/attribute/multienumattribute.h
@@ -2,10 +2,10 @@
 
 #pragma once
 
-#include "multivalueattribute.h"
-#include "enumstorebase.h"
+#include "i_enum_store.h"
 #include "loadedenumvalue.h"
 #include "multivalue.h"
+#include "multivalueattribute.h"
 #include "no_loaded_vector.h"
 
 namespace search {
