+++ searchlib/src/vespa/searchlib/attribute/postingchange.cpp
@@ -102,7 +102,7 @@ removeDupRemovals(std::vector<uint32_t> &removals)
 }
 
 IEnumStore::Index
-EnumIndexMapper::map(IEnumStore::Index original, const EnumStoreComparator & compare) const
+EnumIndexMapper::map(IEnumStore::Index original, const datastore::EntryComparator& compare) const
 {
     (void) compare;
     return original;
@@ -170,14 +170,14 @@ public:
                  const WeightedIndex * entriesOld, size_t szOld,
                  AlwaysWeightedIndexVector & added, AlwaysWeightedIndexVector & changed, AlwaysWeightedIndexVector & removed);
 
-    ActualChangeComputer(const EnumStoreComparator &compare, const EnumIndexMapper &mapper);
+    ActualChangeComputer(const datastore::EntryComparator& compare, const EnumIndexMapper& mapper);
     ~ActualChangeComputer();
 
 private:
     WeightedIndexVector _oldEntries;
     WeightedIndexVector _newEntries;
     vespalib::hash_map<uint32_t, uint32_t> _cachedMapping;
-    const EnumStoreComparator &_compare;
+    const datastore::EntryComparator &_compare;
     const EnumIndexMapper &_mapper;
     const bool _hasFold;
 
@@ -262,7 +262,7 @@ public:
 };
 
 template <typename WeightedIndex>
-ActualChangeComputer<WeightedIndex>::ActualChangeComputer(const EnumStoreComparator &compare,
+ActualChangeComputer<WeightedIndex>::ActualChangeComputer(const datastore::EntryComparator &compare,
                                                           const EnumIndexMapper &mapper)
     : _oldEntries(),
       _newEntries(),
@@ -318,7 +318,7 @@ template <typename MultivalueMapping>
 PostingMap
 PostingChangeComputerT<WeightedIndex, PostingMap>::
 compute(const MultivalueMapping & mvm, const DocIndices & docIndices,
-        const EnumStoreComparator & compare, const EnumIndexMapper & mapper)
+        const datastore::EntryComparator & compare, const EnumIndexMapper & mapper)
 {
     typedef ActualChangeComputer<WeightedIndex> AC;
     AC actualChange(compare, mapper);
@@ -362,13 +362,13 @@ typedef std::vector<std::pair<uint32_t, std::vector<ValueIndex>>> DocIndicesValu
 template WeightedPostingChangeMap PostingChangeComputerT<WeightedIndex, WeightedPostingChangeMap>
              ::compute<WeightedMultiValueMapping>(const WeightedMultiValueMapping &,
                                                    const DocIndicesWeighted &,
-                                                   const EnumStoreComparator &,
+                                                   const datastore::EntryComparator &,
                                                    const EnumIndexMapper &);
 
 template WeightedPostingChangeMap PostingChangeComputerT<ValueIndex, WeightedPostingChangeMap>
              ::compute<ValueMultiValueMapping>(const ValueMultiValueMapping &,
                                                 const DocIndicesValue &,
-                                                const EnumStoreComparator &,
+                                                const datastore::EntryComparator &,
                                                 const EnumIndexMapper &);
 
 }
